chore(jira): rename webhook files to kebab-case
